move interface and abstract Effective spec classes to oracle.kubernetes.operator.processing package
