+++ operator/src/main/java/oracle/kubernetes/operator/processing/EffectiveAdminServerSpec.java
@@ -1,7 +1,9 @@
 // Copyright (c) 2019, 2021, Oracle and/or its affiliates.
 // Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
 
-package oracle.kubernetes.weblogic.domain.model;
+package oracle.kubernetes.operator.processing;
+
+import oracle.kubernetes.weblogic.domain.model.AdminService;
 
 /** Represents the effective configuration for an admin server, as seen by the operator runtime. */
 public interface EffectiveAdminServerSpec extends EffectiveServerSpec {
