+++ operator/src/main/java/oracle/kubernetes/operator/processing/EffectiveClusterSpec.java
@@ -1,7 +1,7 @@
 // Copyright (c) 2018, 2022, Oracle and/or its affiliates.
 // Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
 
-package oracle.kubernetes.weblogic.domain.model;
+package oracle.kubernetes.operator.processing;
 
 import java.util.Collections;
 import java.util.List;
@@ -11,6 +11,7 @@ import javax.annotation.Nonnull;
 import io.kubernetes.client.openapi.models.V1Container;
 import io.kubernetes.client.openapi.models.V1PodSpec;
 import io.kubernetes.client.openapi.models.V1ServiceSpec;
+import oracle.kubernetes.weblogic.domain.model.Shutdown;
 
 public abstract class EffectiveClusterSpec {
