+++ operator/src/main/java/oracle/kubernetes/weblogic/domain/model/EffectiveAdminServerSpecCommonImpl.java
@@ -3,6 +3,7 @@
 
 package oracle.kubernetes.weblogic.domain.model;
 
+import oracle.kubernetes.operator.processing.EffectiveAdminServerSpec;
 import org.apache.commons.lang3.builder.EqualsBuilder;
 import org.apache.commons.lang3.builder.HashCodeBuilder;
 import org.apache.commons.lang3.builder.ToStringBuilder;
