+++ operator/src/main/java/oracle/kubernetes/weblogic/domain/model/EffectiveClusterSpecCommonImpl.java
@@ -9,6 +9,7 @@ import java.util.Map;
 import io.kubernetes.client.openapi.models.V1Container;
 import io.kubernetes.client.openapi.models.V1PodSpec;
 import io.kubernetes.client.openapi.models.V1ServiceSpec;
+import oracle.kubernetes.operator.processing.EffectiveClusterSpec;
 
 public class EffectiveClusterSpecCommonImpl extends EffectiveClusterSpec {
   private final ClusterSpec clusterSpec;
