+++ operator/src/test/java/oracle/kubernetes/weblogic/domain/model/DomainV2Test.java
@@ -27,6 +27,7 @@ import oracle.kubernetes.operator.OverrideDistributionStrategy;
 import oracle.kubernetes.operator.ServerStartPolicy;
 import oracle.kubernetes.operator.ServerStartState;
 import oracle.kubernetes.operator.ShutdownType;
+import oracle.kubernetes.operator.processing.EffectiveServerSpec;
 import org.hamcrest.Matcher;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
